Changed projects page, added in link and highlights.
Added in links to the source projects on github if I specified them, to show that they are interactable they change color when hovering over the items.
Since I was working on hovering items I also made the menu bar show a line under the item that the mouse was currently hovering above.