Worked out the build issue
Travis-CI was running the server and this process never quits. Therefore it was neccessary to rethink the build process, right now I added 'dev' to the test script. The script now checks each page once, if the exit code from those pages comes down to 200 this means all is good, if not the server will throw an error. This seems to work fine except for the fact that the script still does not exit by itself. So I took care of that by including a procces.exit(0) statement when all the tests are done. If this command gets executed it means that all pages finished loading without giving an error.