Worked on the general layout of the site
Made some changes to the css, I moved the css of the pages to a specific stylesheet to keep everything organized.
I changed the menu so that it is now including the header image. This looks a lot better. The links in the menu now contain a little bar beneath them that color's bleu when that item is selected.
Updated projects page to include a development status and sorted them in to two columns by actually defining to medium-6 columns instead of making each item a medium-6 column, this makes sure that all the items are shown below eachother without any whitespace because the item in the other column is bigger. This looks better but required some coding. I used a mixin in Jade to make this as clean as possible.