Changed the projects page's labels
The labels on the projects page are now arrays in the .json file and are processed as an array in the .jade file where we go through each label and use its css value to determine the css and use the text variable to determine the text.
Added in a slight shadow effect for each project.