Made database reading work
You are now able to read data from a database, when read a callback will occur where you are able to do something with the data.
For now I made sure the movies page works, it shows all the movie titles that are in the database. It uses callbacks and thus minmizes the time needed for the page to load.