The menu is no longer a child of the header-image.
This was needed because I wanted the page items to slightly overlap with the header image. This was possible by setting the z-index of the header image to -1. This however made it impossible to press the buttons on the menu, they would not respond. The solution was to make the menu's z-index higher. Since it was a child of the header image I could not figure out how to do this and thus I removed it from its parent the header image. Now the menu/top-bar has been placed at the absolute top of the page.